+++ src/backend/resolvers/mutations/customData/pipeDelimitedScript/index.js
@@ -1,0 +1,10 @@
+const mongoQuery = require('./mongo-query')
+
+const pipeDelimitedScript = async (
+  parent,
+  body,
+  { pulseDevDb },
+  info
+) => mongoQuery(pulseDevDb)
+
+module.exports = pipeDelimitedScript