+++ src/backend/resolvers/mutations/customData/pipeDelimitedScript/numericize-historical-collection.js
@@ -1,0 +1,18 @@
+const convertMongoCollectionFieldsToNumber = db => {
+  const collectionNames = [
+    'payerHistoricalAdditionalCriteria',
+    'payerHistoricalDrgNationalLives',
+    'payerHistoricalPolicyLinks',
+    'payerHistoricalDrgStateLives'
+  ]
+
+  collectionNames.forEach(collectionName => {
+    const cursor = db.collection(collectionName).find()
+
+    cursor.forEach(doc => {
+      doc.year = Number(doc.year)
+      doc.month = Number(doc.month)
+      db.collection(collectionName).save(doc)
+    })
+  })
+}