+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - Started: 12-23-19
 
 ### Added
+- `Polaris Backend`: moved entire `pipe-delimited-script` into `Polaris
 - `Polaris Frontend`: panel item style now supports pseudo-classes, such as `:hover`
 - `Orion`: oncologists column in query tool
 - `Phoenix`: users can now toggle on/off all treatment plans in the resources modal